Move ModuleFinder components to /modules
