+++ website/src/views/modules/ModuleFinderItem.scss
