+++ website/src/views/modules/ModuleFinderPager.scss
